[CF] moved the last of the attached properties out
